refactor(Count.java) : 원시값 포장
