+++ src/main/java/common/vo/Count.java
@@ -1,4 +1,4 @@
-package lotto.winning.domain;
+package common.vo;
 
 public class Count {
