+++ src/main/java/lotto/winning/domain/WinningMoney.java
@@ -1,26 +1,20 @@
 package lotto.winning.domain;
 
+import common.vo.Count;
+
 import static lotto.winning.domain.WinningMoneyType.*;
 
 public class WinningMoney {
 
-    private final int matchCount;
-    private final int lottoCount;
+    private final Count matchCount;
+    private final Count lottoCount;
 
-    public WinningMoney(int matchCount, int lottoCount) {
+    public WinningMoney(Count matchCount, Count lottoCount) {
         this.matchCount = matchCount;
         this.lottoCount = lottoCount;
     }
 
-    public int getLottoCount() {
-        return this.lottoCount;
-    }
-
-    public int getMatchCount() {
-        return this.matchCount;
-    }
-
     public long calculate() {
-        return find(this.matchCount).getMoney() * lottoCount;
+        return find(this.matchCount.getNumber()).getMoney() * lottoCount.getNumber();
     }
 }
